# swiss: implement extendible hashing