--- a/map_test.go
+++ b/map_test.go
@@ -16,14 +16,19 @@ package swiss
 
 import (
 	"fmt"
+	"math"
 	"math/rand"
 	"sort"
 	"testing"
+	"time"
 	"unsafe"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+// TODO(peter): Add metamorphic tests that cross-check behavior at various
+// bucket sizes.
 
 // toBuiltinMap returns the elements as a map[K]V. Useful for testing.
 func (m *Map[K, V]) toBuiltinMap() map[K]V {
@@ -36,19 +41,17 @@ func (m *Map[K, V]) toBuiltinMap() map[K]V {
 }
 
 // TODO(peter): Extracting a random element might be generally useful. Should
-// this be promoted to the public API?
+// this be promoted to the public API? Note that the elements are not selected
+// uniformly randomly. If we promote this method to the public API it should
+// take a rand.Rand.
 func (m *Map[K, V]) randElement() (key K, value V, ok bool) {
-	if m.capacity > 0 {
-		offset := uintptr(rand.Intn(int(m.capacity)))
-		for i := uintptr(0); i <= m.capacity; i++ {
-			j := (i + offset) & m.capacity
-			if (*m.ctrls.At(j) & ctrlEmpty) != ctrlEmpty {
-				s := m.slots.At(j)
-				return s.key, s.value, true
-			}
-		}
-	}
-	return key, value, false
+	// Rely on random iteration order to give us a random element.
+	m.All(func(k K, v V) bool {
+		key, value = k, v
+		ok = true
+		return false
+	})
+	return
 }
 
 func TestLittleEndian(t *testing.T) {
@@ -212,7 +215,7 @@ func bitsetFromString(t *testing.T, str string) bitset {
 }
 
 func TestWasNeverFull(t *testing.T) {
-	m := &Map[int, int]{
+	b := &bucket[int, int]{
 		capacity: 15,
 		ctrls:    makeCtrlBytes(make([]ctrl, 16)),
 	}
@@ -244,145 +247,187 @@ func TestWasNeverFull(t *testing.T) {
 	for _, c := range testCases {
 		t.Run("", func(t *testing.T) {
 			for i := uintptr(0); i < 16; i++ {
-				*m.ctrls.At(i) = 0
+				*b.ctrls.At(i) = 0
 			}
 			for _, i := range c.emptyIndexes {
-				*m.ctrls.At(i) = ctrlEmpty
+				*b.ctrls.At(i) = ctrlEmpty
 			}
-			require.Equal(t, c.expected, m.wasNeverFull(0))
+			require.Equal(t, c.expected, b.wasNeverFull(0))
 		})
 	}
 }
 
 func TestInitialCapacity(t *testing.T) {
 	testCases := []struct {
-		initialCapacity  int
-		expectedCapacity int
+		initialCapacity   int
+		maxBucketCapacity uintptr
+		expectedCapacity  int
+		expectedBuckets   uintptr
 	}{
-		{0, 0},
-		{1, 7},
-		{7, 7},
-		{8, 15},
-		{1000, 1023},
-		{1024, 2047},
+		{0, defaultMaxBucketCapacity, 0, 1},
+		{1, defaultMaxBucketCapacity, 7, 1},
+		{7, defaultMaxBucketCapacity, 7, 1},
+		{8, defaultMaxBucketCapacity, 15, 1},
+		{896, defaultMaxBucketCapacity, 1023, 1},
+		{897, defaultMaxBucketCapacity, 2047, 1},
+		{16, 7, 7 * 4, 4},
+		{65536, 4095, 4095 * 32, 32},
 	}
 	for _, c := range testCases {
 		t.Run("", func(t *testing.T) {
-			m := New[int, int](c.initialCapacity)
-			require.EqualValues(t, c.expectedCapacity, m.capacity)
+			m := New[int, int](c.initialCapacity,
+				WithMaxBucketCapacity[int, int](c.maxBucketCapacity))
+			require.EqualValues(t, c.expectedBuckets, m.bucketCount())
+			require.EqualValues(t, c.expectedCapacity, m.capacity())
 		})
 	}
 }
 
 func TestBasic(t *testing.T) {
-	goodHash := New[int, int](0)
-	badHash := New[int, int](0, WithHash[int, int](func(key *int, seed uintptr) uintptr {
-		return 0
-	}))
+	test := func(t *testing.T, m *Map[int, int]) {
+		const count = 100
 
-	for _, m := range []*Map[int, int]{goodHash, badHash} {
-		t.Run("", func(t *testing.T) {
-			e := make(map[int]int)
-			require.EqualValues(t, 0, m.Len())
-			require.EqualValues(t, 0, m.growthLeft)
+		e := make(map[int]int)
+		require.EqualValues(t, 0, m.Len())
+		require.EqualValues(t, 0, m.bucket0.growthLeft)
 
-			// Non-existent.
-			for i := 0; i < 10; i++ {
-				_, ok := m.Get(i)
-				require.False(t, ok)
-			}
+		// Non-existent.
+		for i := 0; i < count; i++ {
+			_, ok := m.Get(i)
+			require.False(t, ok)
+		}
 
-			// Insert.
-			for i := 0; i < 10; i++ {
-				m.Put(i, i+10)
-				e[i] = i + 10
-				v, ok := m.Get(i)
-				require.True(t, ok)
-				require.EqualValues(t, i+10, v)
-				require.EqualValues(t, i+1, m.Len())
-				require.Equal(t, e, m.toBuiltinMap())
-			}
+		// Insert.
+		for i := 0; i < count; i++ {
+			m.Put(i, i+count)
+			e[i] = i + count
+			v, ok := m.Get(i)
+			require.True(t, ok)
+			require.EqualValues(t, i+count, v)
+			require.EqualValues(t, i+1, m.Len())
+			require.Equal(t, e, m.toBuiltinMap())
+		}
 
-			// Update.
-			for i := 0; i < 10; i++ {
-				m.Put(i, i+20)
-				e[i] = i + 20
-				v, ok := m.Get(i)
-				require.True(t, ok)
-				require.EqualValues(t, i+20, v)
-				require.EqualValues(t, 10, m.Len())
-				require.Equal(t, e, m.toBuiltinMap())
-			}
+		// Update.
+		for i := 0; i < count; i++ {
+			m.Put(i, i+2*count)
+			e[i] = i + 2*count
+			v, ok := m.Get(i)
+			require.True(t, ok)
+			require.EqualValues(t, i+2*count, v)
+			require.EqualValues(t, count, m.Len())
+			require.Equal(t, e, m.toBuiltinMap())
+		}
 
-			// Delete.
-			for i := 0; i < 10; i++ {
-				m.Delete(i)
-				delete(e, i)
-				require.EqualValues(t, 10-i-1, m.Len())
-				_, ok := m.Get(i)
-				require.False(t, ok)
-				require.Equal(t, e, m.toBuiltinMap())
-			}
-		})
+		// Delete.
+		for i := 0; i < count; i++ {
+			m.Delete(i)
+			delete(e, i)
+			require.EqualValues(t, count-i-1, m.Len())
+			_, ok := m.Get(i)
+			require.False(t, ok)
+			require.Equal(t, e, m.toBuiltinMap())
+		}
 	}
+
+	t.Run("normal", func(t *testing.T) {
+		test(t, New[int, int](0))
+	})
+
+	t.Run("degenerate", func(t *testing.T) {
+		testDegenerate := func(t *testing.T, h uintptr) {
+			m := New[int, int](0,
+				WithHash[int, int](func(key *int, seed uintptr) uintptr {
+					return h
+				}),
+				WithMaxBucketCapacity[int, int](7))
+			test(t, m)
+		}
+
+		for _, v := range []uintptr{0, ^uintptr(0)} {
+			t.Run(fmt.Sprintf("%016x", v), func(t *testing.T) {
+				testDegenerate(t, v)
+			})
+		}
+		for i := 0; i < 10; i++ {
+			v := uintptr(rand.Uint64())
+			t.Run(fmt.Sprintf("%016x", v), func(t *testing.T) {
+				testDegenerate(t, v)
+			})
+		}
+	})
 }
 
 func TestRandom(t *testing.T) {
-	goodHash := New[int, int](0)
-	badHash := New[int, int](0, WithHash[int, int](func(key *int, seed uintptr) uintptr {
-		return 0
-	}))
-
-	for _, m := range []*Map[int, int]{goodHash, badHash} {
-		t.Run("", func(t *testing.T) {
-			e := make(map[int]int)
-			for i := 0; i < 10000; i++ {
-				switch r := rand.Float64(); {
-				case r < 0.5: // 50% inserts
-					k, v := rand.Int(), rand.Int()
+	test := func(t *testing.T, m *Map[int, int]) {
+		e := make(map[int]int)
+		for i := 0; i < 10000; i++ {
+			switch r := rand.Float64(); {
+			case r < 0.5: // 50% inserts
+				k, v := rand.Int(), rand.Int()
+				if debug {
+					fmt.Printf("insert %d: %d\n", k, v)
+				}
+				m.Put(k, v)
+				e[k] = v
+			case r < 0.65: // 15% updates
+				if k, _, ok := m.randElement(); !ok {
+					require.EqualValues(t, 0, m.Len(), e)
+				} else {
+					v := rand.Int()
 					if debug {
-						fmt.Printf("insert %d: %d\n", k, v)
+						fmt.Printf("update %d: %d\n", k, v)
 					}
 					m.Put(k, v)
 					e[k] = v
-				case r < 0.65: // 15% updates
-					if k, _, ok := m.randElement(); !ok {
-						require.EqualValues(t, 0, m.Len(), e)
-					} else {
-						v := rand.Int()
-						if debug {
-							fmt.Printf("update %d: %d\n", k, v)
-						}
-						m.Put(k, v)
-						e[k] = v
-					}
-				case r < 0.80: // 15% deletes
-					if k, _, ok := m.randElement(); !ok {
-						require.EqualValues(t, 0, m.Len(), e)
-					} else {
-						if debug {
-							fmt.Printf("delete %d\n", k)
-						}
-						m.Delete(k)
-						delete(e, k)
-					}
-				case r < 0.95: // 25% lookups
-					if k, v, ok := m.randElement(); !ok {
-						require.EqualValues(t, 0, m.Len(), e)
-					} else {
-						if debug {
-							fmt.Printf("lookup %d: %d vs %d\n", k, e[k], v)
-						}
-						require.EqualValues(t, e[k], v)
-					}
-				default: // 5% rehash in place and iterate
-					m.rehashInPlace()
-					require.Equal(t, e, m.toBuiltinMap())
 				}
-				require.EqualValues(t, len(e), m.Len())
+			case r < 0.80: // 15% deletes
+				if k, _, ok := m.randElement(); !ok {
+					require.EqualValues(t, 0, m.Len(), e)
+				} else {
+					if debug {
+						fmt.Printf("delete %d\n", k)
+					}
+					m.Delete(k)
+					delete(e, k)
+				}
+			case r < 0.95: // 25% lookups
+				if k, v, ok := m.randElement(); !ok {
+					require.EqualValues(t, 0, m.Len(), e)
+				} else {
+					if debug {
+						fmt.Printf("lookup %d: %d vs %d\n", k, e[k], v)
+					}
+					require.EqualValues(t, e[k], v)
+				}
+			default: // 5% rehash in place and iterate
+				m.bucket0.rehashInPlace(m)
+				require.Equal(t, e, m.toBuiltinMap())
 			}
-		})
+			require.EqualValues(t, len(e), m.Len())
+		}
 	}
+
+	t.Run("normal", func(t *testing.T) {
+		test(t, New[int, int](0))
+	})
+
+	t.Run("degenerate", func(t *testing.T) {
+		testDegenerate := func(t *testing.T, h uintptr) {
+			m := New[int, int](0,
+				WithHash[int, int](func(key *int, seed uintptr) uintptr {
+					return h
+				}),
+				WithMaxBucketCapacity[int, int](512))
+			test(t, m)
+		}
+
+		for _, v := range []uintptr{0, ^uintptr(0)} {
+			t.Run(fmt.Sprintf("%016x", v), func(t *testing.T) {
+				testDegenerate(t, v)
+			})
+		}
+	})
 }
 
 func TestIterateMutate(t *testing.T) {
@@ -400,7 +445,7 @@ func TestIterateMutate(t *testing.T) {
 	vals := make(map[int]int)
 	m.All(func(k, v int) bool {
 		if (k % 10) == 0 {
-			m.resize(2*m.capacity + 1)
+			m.bucket0.resize(m, 2*m.bucket0.capacity+1)
 		}
 		vals[k] = v
 		return true
@@ -435,7 +480,8 @@ func (a *countingAllocator[K, V]) FreeControls(_ []uint8) {
 
 func TestAllocator(t *testing.T) {
 	a := &countingAllocator[int, int]{}
-	m := New[int, int](0, WithAllocator[int, int](a))
+	m := New[int, int](0, WithAllocator[int, int](a),
+		WithMaxBucketCapacity[int, int](math.MaxUint64))
 
 	for i := 0; i < 100; i++ {
 		m.Put(i, i)
@@ -452,6 +498,35 @@ func TestAllocator(t *testing.T) {
 
 	require.EqualValues(t, expected, a.freeSlots)
 	require.EqualValues(t, expected, a.freeCtrls)
+}
+
+func TestResizeVsSplit(t *testing.T) {
+	if invariants {
+		t.Skip("skipped due to slowness under invariants")
+	}
+
+	count := 1_000_000 + rand.Intn(500_000)
+	m := New[int, int](count, WithMaxBucketCapacity[int, int](0))
+	for i, x := 0, 0; i < count; i++ {
+		x += rand.Intn(128) + 1
+		m.Put(x, x)
+	}
+	start := time.Now()
+	m.bucket0.split(m)
+	if testing.Verbose() {
+		fmt.Printf(" split(%d): %6.3fms\n", count, time.Since(start).Seconds()*1000)
+	}
+
+	m = New[int, int](count, WithMaxBucketCapacity[int, int](math.MaxUint64))
+	for i, x := 0, 0; i < count; i++ {
+		x += rand.Intn(128) + 1
+		m.Put(x, x)
+	}
+	start = time.Now()
+	m.bucket0.resize(m, 2*m.bucket0.capacity+1)
+	if testing.Verbose() {
+		fmt.Printf("resize(%d): %6.3fms\n", count, time.Since(start).Seconds()*1000)
+	}
 }
 
 func BenchmarkStringMap(b *testing.B) {
@@ -585,7 +660,8 @@ func benchmarkSwissMap[K comparable](b *testing.B, keys []K) {
 	n := uint32(len(keys))
 
 	b.Run("Get", func(b *testing.B) {
-		m := New[K, K](len(keys))
+		m := New[K, K](len(keys), WithMaxBucketCapacity[K, K](0))
+		// m := New[K, K](len(keys), WithMaxBucketCapacity[K, K](math.MaxUint64))
 		for _, k := range keys {
 			m.Put(k, k)
 		}
@@ -598,13 +674,15 @@ func benchmarkSwissMap[K comparable](b *testing.B, keys []K) {
 		b.StopTimer()
 
 		assert.True(b, ok)
-		b.ReportMetric(float64(m.Len())/float64(m.capacity), "load-factor")
+		b.ReportMetric(float64(m.Len())/float64(m.capacity()), "load-factor")
+		b.ReportMetric(float64(m.bucketCount()), "buckets")
 	})
 
 	// Rather than benchmark puts and deletes separately, we benchmark them
 	// together so the map stays a constant size.
 	b.Run("PutDelete", func(b *testing.B) {
 		m := New[K, K](len(keys))
+		// m := New[K, K](len(keys), WithMaxBucketCapacity[K, K](math.MaxUint64))
 		for _, k := range keys {
 			m.Put(k, k)
 		}
@@ -617,6 +695,7 @@ func benchmarkSwissMap[K comparable](b *testing.B, keys []K) {
 		}
 		b.StopTimer()
 
-		b.ReportMetric(float64(m.Len())/float64(m.capacity), "load-factor")
+		b.ReportMetric(float64(m.Len())/float64(m.capacity()), "load-factor")
+		b.ReportMetric(float64(m.bucketCount()), "buckets")
 	})
 }