Add extendible hashing on top of the base Swiss Tables design. When there is
only a single bucket, performance is essentially unchanged. The differences
below are likely just noise. The benefit of extendible hashing is to reduce
the maximum operation time when resizing needs to take place. With extendible
hashing resizing is incremental as we can split a single bucket at a time.
With a max bucket capacity of 4095 (the default) splitting of a bucket takes
in the 100us range. Resizing a bucket has a similar cost, but since the max
bucket size is unbound without extendible hashing it is possible to get up to
10s of milliseconds for the resize operation (e.g. for a million element map).
See TestResizeVsSplit for evidence of this.

```
name                                              old time/op      new time/op      delta
StringMap/avgLoad,n=10/swissMap/Get-10                8.40ns ± 0%      8.30ns ± 0%   -1.09%  (p=0.000 n=10+8)
StringMap/avgLoad,n=83/swissMap/Get-10                8.64ns ± 4%      9.12ns ±15%     ~     (p=0.278 n=9+10)
StringMap/avgLoad,n=671/swissMap/Get-10               9.00ns ± 4%      8.91ns ± 3%     ~     (p=0.247 n=10+10)
StringMap/avgLoad,n=5375/swissMap/Get-10              9.35ns ± 1%      9.28ns ± 1%     ~     (p=0.075 n=10+10)
StringMap/avgLoad,n=86015/swissMap/Get-10             10.9ns ± 1%      10.9ns ± 0%   -0.65%  (p=0.000 n=10+9)
Int64Map/avgLoad,n=10/swissMap/Get-10                 4.85ns ± 1%      4.78ns ± 1%   -1.49%  (p=0.000 n=10+10)
Int64Map/avgLoad,n=83/swissMap/Get-10                 5.23ns ±15%      4.88ns ± 9%   -6.77%  (p=0.013 n=10+9)
Int64Map/avgLoad,n=671/swissMap/Get-10                5.42ns ± 3%      5.28ns ± 5%     ~     (p=0.076 n=9+10)
Int64Map/avgLoad,n=5375/swissMap/Get-10               5.68ns ± 1%      5.60ns ± 2%   -1.46%  (p=0.011 n=10+10)
Int64Map/avgLoad,n=86015/swissMap/Get-10              6.88ns ± 0%      6.67ns ± 0%   -3.12%  (p=0.000 n=10+10)
StringMap/avgLoad,n=10/swissMap/PutDelete-10          25.8ns ±32%      27.0ns ±21%     ~     (p=0.218 n=10+10)
StringMap/avgLoad,n=83/swissMap/PutDelete-10          31.8ns ±10%      31.0ns ±12%     ~     (p=0.393 n=10+10)
StringMap/avgLoad,n=671/swissMap/PutDelete-10         34.1ns ± 4%      34.6ns ± 7%     ~     (p=0.436 n=10+10)
StringMap/avgLoad,n=5375/swissMap/PutDelete-10        35.3ns ± 1%      36.1ns ± 4%   +2.03%  (p=0.005 n=10+10)
StringMap/avgLoad,n=86015/swissMap/PutDelete-10       39.5ns ± 1%      40.1ns ± 1%   +1.50%  (p=0.000 n=9+10)
Int64Map/avgLoad,n=10/swissMap/PutDelete-10           19.4ns ±15%      16.9ns ±23%  -12.91%  (p=0.035 n=9+10)
Int64Map/avgLoad,n=83/swissMap/PutDelete-10           21.3ns ± 6%      20.3ns ± 8%   -4.62%  (p=0.041 n=9+10)
Int64Map/avgLoad,n=671/swissMap/PutDelete-10          24.7ns ± 9%      24.5ns ± 7%     ~     (p=0.739 n=10+10)
Int64Map/avgLoad,n=5375/swissMap/PutDelete-10         27.7ns ± 4%      27.5ns ± 2%     ~     (p=0.517 n=10+10)
Int64Map/avgLoad,n=86015/swissMap/PutDelete-10        32.0ns ± 1%      31.5ns ± 1%   -1.76%  (p=0.000 n=10+10)
```

When configured with a maximum bucket capacity of 4095 performance degrades
somewhat when there are multiple buckets:

```
name                                              old time/op      new time/op      delta
StringMap/avgLoad,n=10/swissMap/Get-10                8.30ns ± 0%      8.31ns ± 0%       ~     (p=0.263 n=8+9)
StringMap/avgLoad,n=83/swissMap/Get-10                9.12ns ±15%      8.86ns ±10%       ~     (p=0.564 n=10+10)
StringMap/avgLoad,n=671/swissMap/Get-10               8.91ns ± 3%      8.92ns ± 3%       ~     (p=0.853 n=10+10)
StringMap/avgLoad,n=5375/swissMap/Get-10              9.28ns ± 1%     11.37ns ± 1%    +22.54%  (p=0.000 n=10+9)
StringMap/avgLoad,n=86015/swissMap/Get-10             10.9ns ± 0%      12.5ns ± 1%    +14.90%  (p=0.000 n=9+9)
Int64Map/avgLoad,n=10/swissMap/Get-10                 4.78ns ± 1%      4.78ns ± 1%       ~     (p=0.704 n=10+9)
Int64Map/avgLoad,n=83/swissMap/Get-10                 4.88ns ± 9%      5.26ns ±17%     +7.71%  (p=0.043 n=9+10)
Int64Map/avgLoad,n=671/swissMap/Get-10                5.28ns ± 5%      5.39ns ± 7%       ~     (p=0.353 n=10+10)
Int64Map/avgLoad,n=5375/swissMap/Get-10               5.60ns ± 2%      6.67ns ± 2%    +19.23%  (p=0.000 n=10+9)
Int64Map/avgLoad,n=86015/swissMap/Get-10              6.67ns ± 0%      8.05ns ± 0%    +20.67%  (p=0.000 n=10+8)
StringMap/avgLoad,n=10/swissMap/PutDelete-10          27.0ns ±21%      28.0ns ±21%       ~     (p=0.497 n=10+9)
StringMap/avgLoad,n=83/swissMap/PutDelete-10          31.0ns ±12%      33.9ns ± 8%     +9.52%  (p=0.003 n=10+10)
StringMap/avgLoad,n=671/swissMap/PutDelete-10         34.6ns ± 7%      35.2ns ± 5%       ~     (p=0.280 n=10+10)
StringMap/avgLoad,n=5375/swissMap/PutDelete-10        36.1ns ± 4%      41.2ns ± 3%    +14.34%  (p=0.000 n=10+10)
StringMap/avgLoad,n=86015/swissMap/PutDelete-10       40.1ns ± 1%      45.7ns ± 1%    +14.01%  (p=0.000 n=10+9)
Int64Map/avgLoad,n=10/swissMap/PutDelete-10           16.9ns ±23%      16.5ns ±13%       ~     (p=0.797 n=10+9)
Int64Map/avgLoad,n=83/swissMap/PutDelete-10           20.3ns ± 8%      20.4ns ± 9%       ~     (p=0.912 n=10+10)
Int64Map/avgLoad,n=671/swissMap/PutDelete-10          24.5ns ± 7%      24.8ns ± 7%       ~     (p=0.353 n=10+10)
Int64Map/avgLoad,n=5375/swissMap/PutDelete-10         27.5ns ± 2%      32.2ns ± 3%    +17.06%  (p=0.000 n=10+10)
Int64Map/avgLoad,n=86015/swissMap/PutDelete-10        31.5ns ± 1%      35.6ns ± 1%    +13.22%  (p=0.000 n=10+8)
```

The size of this degradation is somewhat surprising as locating a bucket
involves a right shift of the hash and indexing into an array of pointers.
Even with this performance degradation, `swiss.Map` significantly outperforms
Go's builtin map:

```
name                                         old time/op  new time/op  delta
StringMap/avgLoad,n=10/Map/Get-10            10.2ns ±17%   8.3ns ± 0%  -18.62%  (p=0.000 n=10+9)
StringMap/avgLoad,n=83/Map/Get-10            10.9ns ± 9%   8.9ns ±10%  -18.97%  (p=0.000 n=10+10)
StringMap/avgLoad,n=671/Map/Get-10           15.4ns ± 5%   8.9ns ± 3%  -42.07%  (p=0.000 n=10+10)
StringMap/avgLoad,n=5375/Map/Get-10          25.8ns ± 1%  11.4ns ± 1%  -56.01%  (p=0.000 n=10+9)
StringMap/avgLoad,n=86015/Map/Get-10         30.2ns ± 1%  12.5ns ± 1%  -58.68%  (p=0.000 n=9+9)
Int64Map/avgLoad,n=10/Map/Get-10             5.00ns ± 0%  4.78ns ± 1%   -4.33%  (p=0.000 n=9+9)
Int64Map/avgLoad,n=83/Map/Get-10             5.19ns ± 2%  5.26ns ±17%     ~     (p=0.353 n=10+10)
Int64Map/avgLoad,n=671/Map/Get-10            6.37ns ±10%  5.39ns ± 7%  -15.39%  (p=0.000 n=10+10)
Int64Map/avgLoad,n=5375/Map/Get-10           17.9ns ± 2%   6.7ns ± 2%  -62.80%  (p=0.000 n=9+9)
Int64Map/avgLoad,n=86015/Map/Get-10          23.8ns ± 0%   8.0ns ± 0%  -66.12%  (p=0.000 n=10+8)
```