@RaduBerinde 2024-02-11
I will have to think about it a bot.

I had a different idea. We have the old and new map and maintain an invariant that all hashes (mod old size) below X are in the new map and the ones above are in the new map. Lookup is no slower because we know in which map to look.

Before every insert we advance X by 4 (moving up to 4 elements to the new map). We insert the new element in either map depending on X. If we start this process before the map is 3/4 full we won't overflow the old map.

I don't know how much of a concern resizing really is in practice in CRDB usecased.