@petermattis 2024-02-12
Another pro for the extendible hashing approach is that if we make the bucket size threshold configurable we can configure a very large bucket size in order to fall back to the current behavior.