@petermattis 2024-02-14
This is ready for a looksie. I broke a test which is easy to fix, and need to add some more tests. Perf is unchanged if we have only a single bucket. If we have more than 1 bucket perf declines a bit, though still much better than runtime map. Here, old is 1 bucket, and new is more than 1 bucket:

```
name                                        old time/op      new time/op      delta
StringMap/avgLoad,n=10/swissMap/Get-10          8.33ns ± 1%      8.36ns ± 1%     +0.37%  (p=0.019 n=8+10)
StringMap/maxLoad,n=13/swissMap/Get-10          8.33ns ± 0%      8.33ns ± 0%       ~     (p=1.000 n=8+8)
StringMap/minLoad,n=14/swissMap/Get-10          8.59ns ± 1%      8.34ns ± 1%     -2.99%  (p=0.000 n=8+8)
StringMap/avgLoad,n=83/swissMap/Get-10          8.89ns ± 8%      8.77ns ±10%       ~     (p=0.615 n=10+10)
StringMap/maxLoad,n=111/swissMap/Get-10         10.1ns ± 9%      10.1ns ± 2%       ~     (p=0.983 n=10+8)
StringMap/minLoad,n=112/swissMap/Get-10         8.54ns ± 0%      8.35ns ± 2%     -2.25%  (p=0.000 n=9+9)
StringMap/avgLoad,n=671/swissMap/Get-10         9.17ns ± 4%      8.94ns ± 2%     -2.51%  (p=0.011 n=10+10)
StringMap/maxLoad,n=895/swissMap/Get-10         10.3ns ± 3%      10.3ns ± 4%       ~     (p=0.986 n=10+10)
StringMap/minLoad,n=896/swissMap/Get-10         8.43ns ± 1%      8.49ns ± 1%       ~     (p=0.142 n=9+9)
StringMap/avgLoad,n=5375/swissMap/Get-10        9.42ns ± 6%     10.82ns ± 1%    +14.92%  (p=0.000 n=10+9)
StringMap/maxLoad,n=7167/swissMap/Get-10        11.1ns ± 1%      10.9ns ± 1%     -1.74%  (p=0.000 n=9+10)
StringMap/minLoad,n=7168/swissMap/Get-10        8.78ns ± 1%     10.88ns ± 0%    +23.91%  (p=0.000 n=9+9)
StringMap/avgLoad,n=86015/swissMap/Get-10       11.1ns ± 1%      15.4ns ± 6%    +39.28%  (p=0.000 n=9+9)
StringMap/maxLoad,n=114687/swissMap/Get-10      13.9ns ± 1%      17.0ns ± 9%    +22.20%  (p=0.000 n=10+9)
StringMap/minLoad,n=114688/swissMap/Get-10      14.4ns ± 3%      16.8ns ± 3%    +16.38%  (p=0.000 n=10+10)

name                                        old buckets      new buckets      delta
StringMap/avgLoad,n=10/swissMap/Get-10            1.00 ± 0%        1.00 ± 0%       ~     (all equal)
StringMap/maxLoad,n=13/swissMap/Get-10            1.00 ± 0%        1.00 ± 0%       ~     (all equal)
StringMap/minLoad,n=14/swissMap/Get-10            1.00 ± 0%        1.00 ± 0%       ~     (all equal)
StringMap/avgLoad,n=83/swissMap/Get-10            1.00 ± 0%        1.00 ± 0%       ~     (all equal)
StringMap/maxLoad,n=111/swissMap/Get-10           1.00 ± 0%        1.00 ± 0%       ~     (all equal)
StringMap/minLoad,n=112/swissMap/Get-10           1.00 ± 0%        1.00 ± 0%       ~     (all equal)
StringMap/avgLoad,n=671/swissMap/Get-10           1.00 ± 0%        1.00 ± 0%       ~     (all equal)
StringMap/maxLoad,n=895/swissMap/Get-10           1.00 ± 0%        1.00 ± 0%       ~     (all equal)
StringMap/minLoad,n=896/swissMap/Get-10           1.00 ± 0%        1.00 ± 0%       ~     (all equal)
StringMap/avgLoad,n=5375/swissMap/Get-10          1.00 ± 0%        4.00 ± 0%   +300.00%  (p=0.000 n=10+10)
StringMap/maxLoad,n=7167/swissMap/Get-10          1.00 ± 0%        4.00 ± 0%   +300.00%  (p=0.000 n=10+10)
StringMap/minLoad,n=7168/swissMap/Get-10          1.00 ± 0%        4.00 ± 0%   +300.00%  (p=0.000 n=10+10)
StringMap/avgLoad,n=86015/swissMap/Get-10         1.00 ± 0%       64.00 ± 0%  +6300.00%  (p=0.000 n=10+10)
StringMap/maxLoad,n=114687/swissMap/Get-10        1.00 ± 0%       64.00 ± 0%  +6300.00%  (p=0.000 n=10+10)
StringMap/minLoad,n=114688/swissMap/Get-10        1.00 ± 0%       64.00 ± 0%  +6300.00%  (p=0.000 n=10+10)
```