@RaduBerinde 2024-02-17
I was thinking - could we make `New` return a `Map` instead of `*Map` so the struct itself can live on the stack (or be part of larger structs)?

I think we would need some changes to the implementation here so we don't make the pointer to the Map escape. For one, we can't point to `&bucket0` anymore from the directory, but we don't need to, we can just shallow copy the bucket to a new bucket. We would also need to rethink `dirEntries` (we could just special case the globalDepth==0 case wherever we use it).