@petermattis 2024-02-18
> I was thinking - could we make New return a Map instead of *Map so the struct itself can live on the stack (or be part of larger structs)?
>
> I think we would need some changes to the implementation here so we don't make the pointer to the Map escape. For one, we can't point to &bucket0 anymore from the directory, but we don't need to, we can just shallow copy the bucket to a new bucket. We would also need to rethink dirEntries (we could just special case the globalDepth==0 case wherever we use it).

I did add support for a `Map.Init` method in https://github.com/cockroachdb/swiss/pull/17 which allows a `Map` to be embedded by value. That doesn't allow the struct to live on the stack by itself, but I'm not sure if that is a common case worth optimizing. Seems like the more common case is embedding a `Map` in another struct. Perhaps I should add a `Make` function as well, similar to `New` but it returns a `Map` instead of a `*Map`. I suspect it is better to just wait for a bit of usage experience.