@RaduBerinde 2024-02-18
Thinking more about it, it's not safe to copy a Map by value so we shouldn't encourage that pattern with a Make function.