@petermattis 2024-02-18
Good point. We can and should protect against that.